requirements.txt
streamlit
requests
customtkinter
pillow